requirements.txt
streamlit>=1.31.0
streamlit-authenticator>=0.2.1
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
cohere>=4.0.0
PyMuPDF>=1.23.0
faiss-cpu>=1.7.4
python-docx>=1.0.0
python-pptx>=0.6.21
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.2
